+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.15.3] - 2026-08-04 - Validated run buttons
+
 ### Changed
 
 - Add `Interface.run_button()` so CLI arguments are fully validated before